requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
httpx==0.28.1
python-dotenv==1.2.1
python-telegram-bot==22.6
huggingface_hub==1.3.7
transformers==5.0.0
tokenizers==0.22.2
numpy==1.26.4
scipy==1.10.1
networkx==3.4.2
PyYAML==6.0.3
pydantic==2.12.5
httpcore==1.0.9
Jinja2==3.1.6
MarkupSafe==2.1.5
safetensors==0.7.0
tqdm==4.67.2
regex==2026.1.15
typing_extensions==4.15.0
pytest